removed-requirements.txt
python-gitlab~=5.1.0
PyGithub~=2.5.0
djangorestframework~=3.15.2
google-cloud-secret-manager~=2.22.0
google-crc32c~=1.6.0
